+++ scripts/macros/classFeatures/cleric/natureDomain/charmAnimalsAndPlants.js
@@ -1,0 +1,6 @@
+import {chris} from '../../../../helperFunctions.js';
+export async function charmAnimalsAndPlants({speaker, actor, token, character, item, args, scope, workflow}) {
+    let validTargets = Array.from(workflow.targets).filter(i => ['beast', 'plant'].includes(chris.raceOrType(i.actor))).map(i => i.id);
+    console.log(validTargets);
+    chris.updateTargets(validTargets);
+}