+++ scripts/macros/classFeatures/monk/focusedAim.js
@@ -2,10 +2,10 @@ import {chris} from '../../../helperFunctions.js';
 import {queue} from '../../../utility/queue.js';
 export async function focusedAim({speaker, actor, token, character, item, args, scope, workflow}) {
     if (workflow.targets.size != 1 || workflow.isFumble) return;
-    let feature = await workflow.actor.items.getName('Ki Points');
+    let feature = chris.getItem(workflow.actor, 'Ki Points');
     if (!feature) return;
     let featureUses = feature.system.uses.value;
-    if (featureUses === 0) return;
+    if (!featureUses) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'focusedAim', 151);
     if (!queueSetup) return;
     let attackTotal = workflow.attackTotal;
@@ -19,7 +19,7 @@ export async function focusedAim({speaker, actor, token, character, item, args, 
     if (featureUses >= 2) featureMenu.push(['Yes (2 Ki / +4 to hit)', 4]);
     if (featureUses >= 3) featureMenu.push(['Yes (3 Ki / +6 to hit)', 6]);
     featureMenu.push(['No', false]);
-    let useFeature = await chris.dialog('Attack roll (' + attackTotal + ') missed.  Use Focused Aim?', featureMenu);
+    let useFeature = await chris.dialog(feature.name, featureMenu, 'Attack roll (' + attackTotal + ') missed.  Use Focused Aim?');
     if (!useFeature) {
         queue.remove(workflow.item.uuid);
         return;
@@ -27,12 +27,7 @@ export async function focusedAim({speaker, actor, token, character, item, args, 
     let updatedRoll = await chris.addToRoll(workflow.attackRoll, useFeature);
     workflow.setAttackRoll(updatedRoll);
     feature.update({'system.uses.value': featureUses - (useFeature / 2)});
-    let effect = chris.findEffect(workflow.actor, 'Focused Aim');
-    if (!effect) {
-        queue.remove(workflow.item.uuid);
-        return;
-    }
-    let originItem = await fromUuid(effect.origin);
-    await originItem.use();
+    let originItem = chris.getItem(workflow.actor, 'Focused Aim');
+    if (originItem) await originItem.use();
     queue.remove(workflow.item.uuid);
 }