+++ scripts/macros/raceFeatures/aasimar/necroticShroud.js
@@ -1,12 +1,11 @@
+import {constants} from '../../../constants.js';
 import {chris} from '../../../helperFunctions.js';
 import {queue} from '../../../utility/queue.js';
 async function attack({speaker, actor, token, character, item, args, scope, workflow}) {
     let pass = args[0].macroPass;
     if (workflow.hitTargets.size === 0) return;
     if (!(pass === 'postDamageRoll' || pass === 'preDamageApplication')) return;
-    let effect = chris.findEffect(workflow.actor, 'Celestial Revelation (Necrotic Shroud)');
-    if (!effect) return;
-    let feature = await fromUuid(effect.origin);
+    let feature = chris.getItem(workflow.actor, 'Celestial Revelation (Necrotic Shroud)')
     if (!feature) return;
     let useFeature = chris.perTurnCheck(feature, 'feature', 'necroticShroud', true, workflow.token.id);
     if (!useFeature) return;
@@ -15,12 +14,12 @@ async function attack({speaker, actor, token, character, item, args, scope, work
             if (workflow.hitTargets.size != 1) return;
             let queueSetup = await queue.setup(workflow.item.uuid, 'necroticShroud', 249);
             if (!queueSetup) return;
-            let selected = await chris.dialog('Celestial Revelation: Add extra damage?', [['Yes', true], ['No', false]]);
+            let selected = await chris.dialog(feature.name, constants.yesNo, 'Add extra damage from ' + feature.name + '?');
             if (!selected) {
                 queue.remove(workflow.item.uuid);
                 return;
             }
-            if (!(game.combat === null || game.combat === undefined)) await feature.setFlag('chris-premades', 'feature.necroticShroud.turn', game.combat.round + '-' + game.combat.turn);
+            await chris.setTurnCheck(feature, 'feature', 'necroticShroud');
             let damageFormula = workflow.damageRoll._formula + ' + ' + workflow.actor.system.attributes.prof + '[necrotic]';
             let damageRoll = await new Roll(damageFormula).roll({async: true});
             await workflow.setDamageRoll(damageRoll);
@@ -30,61 +29,24 @@ async function attack({speaker, actor, token, character, item, args, scope, work
             if (workflow.hitTargets.size <= 1) return;
             let queueSetup2 = queue.setup(workflow.item.uuid, 'necroticShroud', 249);
             if (!queueSetup2) return;
-            let buttons = [
-                {
-                    'label': 'Yes',
-                    'value': true
-                }, {
-                    'label': 'No',
-                    'value': false
-                }
-            ];
-            let selection = await chris.selectTarget('Celestial Revelation: Add extra damage?', buttons, workflow.targets, false, 'one');
+            let selection = await chris.selectTarget('Celestial Revelation: Add extra damage?', constants.yesNoButton, workflow.targets, true, 'one');
             if (selection.buttons === false) {
                 queue.remove(workflow.item.uuid);
                 return;
             }
-            if (!(game.combat === null || game.combat === undefined)) await feature.setFlag('chris-premades', 'feature.necroticShroud.turn', game.combat.round + '-' + game.combat.turn);
-            let targetTokenID = selection.inputs.find(id => id != false);
-            if (!targetTokenID) {
+            await chris.setTurnCheck(feature, 'feature', 'necroticShroud');
+            let targetTokenUuid = selection.inputs.find(i => i);
+            if (!targetTokenUuid) {
                 queue.remove(workflow.item.uuid);
                 return;
             }
-            let targetDamage = workflow.damageList.find(i => i.tokenId === targetTokenID);
-            let targetActor = canvas.scene.tokens.get(targetDamage.tokenId).actor;
-            if (!targetActor) {
+            let targetToken = await fromUuid(targetTokenUuid);
+            if (!targetToken) {
                 queue.remove(workflow.item.uuid);
                 return;
             }
-            if (!(game.combat === null || game.combat === undefined)) await feature.setFlag('chris-premades', 'feature.necroticShroud.turn', currentTurn);
-            let hasDI = chris.checkTrait(targetActor, 'di', 'necrotic');
-            if (hasDI) {
-                queue.remove(workflow.item.uuid);
-                return;
-            }
-            let damageTotal = workflow.actor.system.attributes.prof;
-            let hasDR = chris.checkTrait(targetActor, 'dr', 'necrotic');
-            if (hasDR) damageTotal = Math.floor(damageTotal / 2);
-            targetDamage.damageDetail[0].push(
-                {
-                    'damage': damageTotal,
-                    'type': 'necrotic'
-                }
-            );
-            targetDamage.totalDamage += damageTotal;
-            targetDamage.appliedDamage += damageTotal;
-            targetDamage.hpDamage += damageTotal;
-            if (targetDamage.oldTempHP > 0) {
-                if (targetDamage.oldTempHP >= damageTotal) {
-                    targetDamage.newTempHP -= damageTotal;
-                } else {
-                    let leftHP = damageTotal - targetDamage.oldTempHP;
-                    targetDamage.newTempHP = 0;
-                    targetDamage.newHP -= leftHP;
-                }
-            } else {
-                targetDamage.newHP -= damageTotal;
-            }
+            chris.addDamageDetailDamage(targetToken, workflow.actor.system.attributes.prof, 'necrotic', workflow);
+            await feature.displayCard();
             queue.remove(workflow.item.uuid);
             return;
     }
@@ -101,7 +63,6 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
                 'priority': 20
             }
         ],
-        'transfer': false,
         'origin': workflow.item.uuid,
         'duration': {
             'seconds': 60
@@ -110,7 +71,7 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     await chris.createEffect(workflow.actor, effectData);
 }
 async function turn(origin) {
-    await origin.setFlag('chris-premades', 'feature.necroticShroud.turn', false);
+    await chris.setTurnCheck(origin, 'feature', 'necroticShroud', true);
 }
 export let necroticShroud = {
     'attack': attack,